backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
itsdangerous==2.1.2
python-multipart==0.0.6
alembic==1.12.1
httpx==0.25.2
redis==5.0.1
authlib==1.3.0
email-validator==2.1.0

